Emscripten WebGL API (#11400)
* XXX wip hashing on Emscripten WebGL API

* Custom Emscripten WebGL API

* Move WebGL related API from html5.h to html5_webgl.h. Add Emscripten HTML5 WebGL API.

* Add proxying and testing, and docs

* Remove old dead code

* Remove unneeded include

* Use built-in type

* Add missing html5_webgl.js link directive

* Add documentation about array parameters

* Fix WebGL link via -lwebgl.js

* flake

* Revert glBufferData change

* Include html5.h from html5_webgl.h

* Rename _write_array to
