+++ system/include/emscripten/html5.h
@@ -75,7 +75,6 @@ extern "C" {
 
 #define EMSCRIPTEN_RESULT int
 
-
 #define EMSCRIPTEN_RESULT_SUCCESS              0
 #define EMSCRIPTEN_RESULT_DEFERRED             1
 #define EMSCRIPTEN_RESULT_NOT_SUPPORTED       -1
@@ -427,85 +426,10 @@ extern EMSCRIPTEN_RESULT emscripten_vibrate_pattern(int *msecsArray, int numEntr
 typedef const char *(*em_beforeunload_callback)(int eventType, const void *reserved, void *userData);
 extern EMSCRIPTEN_RESULT emscripten_set_beforeunload_callback_on_thread(void *userData, em_beforeunload_callback callback, pthread_t targetThread);
 
-typedef int EMSCRIPTEN_WEBGL_CONTEXT_HANDLE;
-
-typedef int EMSCRIPTEN_WEBGL_CONTEXT_PROXY_MODE;
-#define EMSCRIPTEN_WEBGL_CONTEXT_PROXY_DISALLOW 0
-#define EMSCRIPTEN_WEBGL_CONTEXT_PROXY_FALLBACK 1
-#define EMSCRIPTEN_WEBGL_CONTEXT_PROXY_ALWAYS   2
-
-typedef int EM_WEBGL_POWER_PREFERENCE;
-#define EM_WEBGL_POWER_PREFERENCE_DEFAULT 0
-#define EM_WEBGL_POWER_PREFERENCE_LOW_POWER 1
-#define EM_WEBGL_POWER_PREFERENCE_HIGH_PERFORMANCE 2
-
-typedef struct EmscriptenWebGLContextAttributes {
-  EM_BOOL alpha;
-  EM_BOOL depth;
-  EM_BOOL stencil;
-  EM_BOOL antialias;
-  EM_BOOL premultipliedAlpha;
-  EM_BOOL preserveDrawingBuffer;
-  EM_WEBGL_POWER_PREFERENCE powerPreference;
-  EM_BOOL failIfMajorPerformanceCaveat;
-
-  int majorVersion;
-  int minorVersion;
-
-  EM_BOOL enableExtensionsByDefault;
-  EM_BOOL explicitSwapControl;
-  EMSCRIPTEN_WEBGL_CONTEXT_PROXY_MODE proxyContextToMainThread;
-  EM_BOOL renderViaOffscreenBackBuffer;
-} EmscriptenWebGLContextAttributes;
-
-extern void emscripten_webgl_init_context_attributes(EmscriptenWebGLContextAttributes *attributes);
-
-extern EMSCRIPTEN_WEBGL_CONTEXT_HANDLE emscripten_webgl_create_context(const char *target, const EmscriptenWebGLContextAttributes *attributes);
-
-extern EMSCRIPTEN_RESULT emscripten_webgl_make_context_current(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context);
-
-extern EMSCRIPTEN_WEBGL_CONTEXT_HANDLE emscripten_webgl_get_current_context(void);
-
-extern EMSCRIPTEN_RESULT emscripten_webgl_get_drawing_buffer_size(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context, int *width, int *height);
-
-extern EMSCRIPTEN_RESULT emscripten_webgl_get_context_attributes(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context, EmscriptenWebGLContextAttributes *outAttributes);
-
-extern EMSCRIPTEN_RESULT emscripten_webgl_destroy_context(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context);
-
-extern EM_BOOL emscripten_webgl_enable_extension(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context, const char *extension);
-
-extern EM_BOOL emscripten_webgl_enable_ANGLE_instanced_arrays(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context);
-
-extern EM_BOOL emscripten_webgl_enable_OES_vertex_array_object(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context);
-
-extern EM_BOOL emscripten_webgl_enable_WEBGL_draw_buffers(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context);
-
-extern EM_BOOL emscripten_webgl_enable_WEBGL_draw_instanced_base_vertex_base_instance(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context);
-
-typedef EM_BOOL (*em_webgl_context_callback)(int eventType, const void *reserved, void *userData);
-extern EMSCRIPTEN_RESULT emscripten_set_webglcontextlost_callback_on_thread(const char *target, void *userData, EM_BOOL useCapture, em_webgl_context_callback callback, pthread_t targetThread);
-extern EMSCRIPTEN_RESULT emscripten_set_webglcontextrestored_callback_on_thread(const char *target, void *userData, EM_BOOL useCapture, em_webgl_context_callback callback, pthread_t targetThread);
-
-extern EM_BOOL emscripten_is_webgl_context_lost(EMSCRIPTEN_WEBGL_CONTEXT_HANDLE context);
-
-extern EMSCRIPTEN_RESULT emscripten_webgl_commit_frame(void);
-
-extern EM_BOOL emscripten_supports_offscreencanvas(void);
-
-// Returns function pointers to WebGL 1 functions. Please avoid using this function ever - all WebGL1/GLES2 functions, even those for WebGL1 extensions, are available to user code via static linking. Calling GL functions
-// via function pointers obtained here is slow, and using this function can greatly increase resulting compiled program size. This functionality is available only for easier program code porting purposes, but be aware
-// that calling this is causing a noticeable performance and compiled code size hit.
-extern void *emscripten_webgl1_get_proc_address(const char *name);
-
-// Returns function pointers to WebGL 2 functions. Please avoid using this function ever - all WebGL2/GLES3 functions, even those for WebGL2 extensions, are available to user code via static linking. Calling GL functions
-// via function pointers obtained here is slow, and using this function can greatly increase resulting compiled program size. This functionality is available only for easier program code porting purposes, but be aware
-// that calling this is causing a noticeable performance and compiled code size hit.
-extern void *emscripten_webgl2_get_proc_address(const char *name);
-
-// Combines emscripten_webgl1_get_proc_address() and emscripten_webgl2_get_proc_address() to return function pointers to both WebGL1 and WebGL2 functions. Same drawbacks apply.
-extern void *emscripten_webgl_get_proc_address(const char *name);
-
+// Sets the canvas.width & canvas.height properties.
 extern EMSCRIPTEN_RESULT emscripten_set_canvas_element_size(const char *target, int width, int height);
+
+// Returns the canvas.width & canvas.height properties.
 extern EMSCRIPTEN_RESULT emscripten_get_canvas_element_size(const char *target, int *width, int *height);
 
 extern EMSCRIPTEN_RESULT emscripten_set_element_css_size(const char *target, double width, double height);
@@ -551,8 +475,6 @@ extern void emscripten_html5_remove_all_event_listeners(void);
 #define emscripten_set_batterychargingchange_callback(userData, callback)                     emscripten_set_batterychargingchange_callback_on_thread(          (userData),               (callback), EM_CALLBACK_THREAD_CONTEXT_CALLING_THREAD)
 #define emscripten_set_batterylevelchange_callback(userData, callback)                        emscripten_set_batterylevelchange_callback_on_thread(             (userData),               (callback), EM_CALLBACK_THREAD_CONTEXT_CALLING_THREAD)
 #define emscripten_set_beforeunload_callback(userData, callback)                              emscripten_set_beforeunload_callback_on_thread(                   (userData),               (callback), EM_CALLBACK_THREAD_CONTEXT_MAIN_BROWSER_THREAD)
-#define emscripten_set_webglcontextlost_callback(target, userData, useCapture, callback)      emscripten_set_webglcontextlost_callback_on_thread(     (target), (userData), (useCapture), (callback), EM_CALLBACK_THREAD_CONTEXT_CALLING_THREAD)
-#define emscripten_set_webglcontextrestored_callback(target, userData, useCapture, callback)  emscripten_set_webglcontextrestored_callback_on_thread( (target), (userData), (useCapture), (callback), EM_CALLBACK_THREAD_CONTEXT_CALLING_THREAD)
 
 extern long emscripten_set_timeout(void (*cb)(void *userData), double msecs, void *userData);
 extern void emscripten_clear_timeout(long setTimeoutId);
@@ -584,3 +506,6 @@ extern void emscripten_unwind_to_js_event_loop(void) __attribute__((noreturn));
 #ifdef __cplusplus
 } // ~extern "C"
 #endif
+
+// For backwards compatibility, always include html5_webgl.h from html5.h
+#include <emscripten/html5_webgl.h>
